feat: refactor device components and improve integration handling
