+++ src/app/devices/details/DeviceDetailsClient.tsx
@@ -9,7 +9,7 @@ import { Card, CardContent, CardHeader, CardTitle, CardDescription } from '@/com
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
 import { ArrowLeft, PlusCircle, RefreshCw, History, SlidersHorizontal, Info, Clock, AlertTriangle, ChevronRight, ChevronLeft, Trash2, Zap } from 'lucide-react';
-import { DeviceIcon, StatusBadge as DeviceStatusBadge, mapApiIconToIconType } from '@/app/devices/page';
+import { DeviceIcon, StatusBadge as DeviceStatusBadge, mapApiIconToIconType } from '@/components/devices/DeviceComponents';
 import { useAuth } from '@/contexts/AuthContext';
 import { format, formatDistanceToNowStrict, parseISO, formatDistanceStrict } from 'date-fns';
 import { cn } from '@/lib/utils';
@@ -36,7 +36,7 @@ import { IdentifierDisplay } from '@/components/shared/IdentifierDisplay';
 interface CertificateHistoryEntry {
   version: string;
   serialNumber: string;
-  apiStatus?: string;
+  apiStatus: string;
   revocationReason?: string;
   revocationTimestamp?: string;
   isSuperseded: boolean;
@@ -267,7 +267,7 @@ export default function DeviceDetailsClient() {
                 return {
                     version: version,
                     serialNumber: certData.serialNumber,
-                    apiStatus: certData.apiStatus,
+                    apiStatus: certData.apiStatus || 'ACTIVE',
                     revocationReason: certData.revocationReason,
                     revocationTimestamp: certData.revocationTimestamp,
                     isSuperseded: isSuperseded,
@@ -277,10 +277,11 @@ export default function DeviceDetailsClient() {
                     validFrom: certData.validFrom,
                     validTo: certData.validTo,
                     lifespan: formatDistanceStrict(parseISO(certData.validTo), parseISO(certData.validFrom)),
-                };
+                } as CertificateHistoryEntry;
             });
 
-            const historyEntries = (await Promise.all(certPromises)).filter((e): e is CertificateHistoryEntry => e !== null);
+            const allEntries = await Promise.all(certPromises);
+            const historyEntries = allEntries.filter((e): e is CertificateHistoryEntry => e !== null);
             setCertificateHistory(historyEntries);
 
         } catch (err: any) {
@@ -343,7 +344,7 @@ export default function DeviceDetailsClient() {
                     const historyEntry: CertificateHistoryEntry = {
                         version: associatedVersion || 'N/A',
                         serialNumber: certData.serialNumber,
-                        apiStatus: certData.apiStatus,
+                        apiStatus: certData.apiStatus || 'ACTIVE',
                         revocationReason: certData.revocationReason,
                         revocationTimestamp: certData.revocationTimestamp,
                         isSuperseded: isSuperseded,
@@ -951,8 +952,6 @@ export default function DeviceDetailsClient() {
         device={device}
         ra={raForIntegration}
         availableIntegrations={availableIntegrations}
-        activeIntegration={activeIntegration}
-        setActiveIntegration={setActiveIntegration}
         isUpdating={isForcingUpdate}
       />
     </div>
