+++ src/components/devices/DeviceComponents.tsx
@@ -1,0 +1,55 @@
+import React from 'react';
+import { Badge } from "@/components/ui/badge";
+import { cn } from '@/lib/utils';
+import { HelpCircle } from "lucide-react";
+import { getLucideIconByName } from '@/components/shared/DeviceIconSelectorModal';
+
+type DeviceStatus = 'ACTIVE' | 'NO_IDENTITY' | 'RENEWAL_PENDING' | 'EXPIRING_SOON' | 'EXPIRED' | 'REVOKED' | 'DECOMMISSIONED';
+
+export const StatusBadge: React.FC<{ status: DeviceStatus }> = ({ status }) => {
+  let badgeClass = "";
+  switch (status) {
+    case 'ACTIVE':
+      badgeClass = "bg-green-100 text-green-700 dark:bg-green-700/30 dark:text-green-300 border-green-300 dark:border-green-700";
+      break;
+    case 'RENEWAL_PENDING':
+      badgeClass = "bg-yellow-100 text-yellow-700 dark:bg-yellow-700/30 dark:text-yellow-300 border-yellow-300 dark:border-yellow-700";
+      break;
+    case 'EXPIRING_SOON':
+      badgeClass = "bg-orange-100 text-orange-700 dark:bg-orange-700/30 dark:text-orange-300 border-orange-300 dark:border-orange-700";
+      break;
+    case 'EXPIRED':
+      badgeClass = "bg-purple-100 text-purple-700 dark:bg-purple-700/30 dark:text-purple-300 border-purple-300 dark:border-purple-700";
+      break;
+    case 'REVOKED':
+      badgeClass = "bg-red-100 text-red-700 dark:bg-red-700/30 dark:text-red-300 border-red-300 dark:border-red-700";
+      break;
+    case 'NO_IDENTITY':
+      badgeClass = "bg-sky-100 text-sky-700 dark:bg-sky-700/30 dark:text-sky-300 border-sky-300 dark:border-sky-700";
+      break;
+    case 'DECOMMISSIONED':
+      badgeClass = "bg-gray-100 text-gray-600 dark:bg-gray-800/30 dark:text-gray-400 border-gray-400 dark:border-gray-600";
+      break;
+    default:
+      badgeClass = "bg-muted text-muted-foreground border-border";
+  }
+  return <Badge variant="outline" className={cn("text-xs capitalize", badgeClass)}>{status.replace('_', ' ').toLowerCase()}</Badge>;
+};
+
+export const mapApiIconToIconType = (apiIcon: string): string => {
+  return apiIcon || 'HelpCircle';
+};
+
+export const DeviceIcon: React.FC<{ type: string; iconColor?: string; bgColor?: string; }> = ({ type, iconColor, bgColor }) => {
+  const IconComponent = getLucideIconByName(type);
+
+  return (
+    <div className={cn("p-1.5 rounded-md inline-flex items-center justify-center")} style={{ backgroundColor: bgColor || '#F0F8FF' }}>
+      {IconComponent ? (
+        <IconComponent className={cn("h-5 w-5")} style={{ color: iconColor || '#0f67ff' }} />
+      ) : (
+        <HelpCircle className={cn("h-5 w-5")} style={{ color: iconColor || '#0f67ff' }} />
+      )}
+    </div>
+  );
+};
